Partially undo fbc7e6b
I'd gotten a little overzealous in applying the pattern needed for the
copy operation.  Turns out it was necessary there because we weren't
doing anything locally first.  When we pulled the
ImapDB.EmailIdentifiers out of the local folder for any local op, we
didn't then need to pass those back in to get the UIDs.  This was
causing move not to work.